+++ babel.config.js
@@ -1,5 +1,6 @@
 module.exports = {
   presets: [
+    '@babel/preset-typescript',
     [
       '@babel/preset-env',
       {
@@ -16,6 +17,7 @@ module.exports = {
     ],
   ],
   plugins: [
+    [ './babel-plugin', { publicPath: true }],
     [ '@babel/plugin-proposal-nullish-coalescing-operator', { loose: true }],
     [ '@babel/plugin-proposal-optional-chaining', { loose: true }],
     [ '@babel/plugin-proposal-object-rest-spread', { useBuiltIns: true }],