+++ src/index.ts
@@ -1,8 +1,11 @@
 import Jpex from './Jpex';
 import { resolve as registerResolveFactory } from './factories';
-import {
+import type {
   JpexInstance,
   SetupConfig,
+  Options,
+  NamedParameters,
+  Resolve,
 } from './types';
 import {
   Lifecycle,
@@ -14,9 +17,15 @@ registerResolveFactory(jpex);
 
 export {
   jpex,
+};
+
+export type {
   JpexInstance,
   SetupConfig,
   Lifecycle,
+  Options,
+  NamedParameters,
+  Resolve,
 };
 
 export default jpex;